requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
pymongo
certifi
pymongo[srv]==3.11
scikit-learn
dill
pyaml
xgboost
lightgbm
-e .